working on norm

+++ srcs/utlis.c
@@ -6,7 +6,7 @@
 /*   By: iharchi <iharchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/28 12:28:57 by iharchi           #+#    #+#             */
-/*   Updated: 2021/06/26 16:26:11 by iharchi          ###   ########.fr       */
+/*   Updated: 2021/06/26 17:44:09 by iharchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,77 +14,77 @@
 
 int ft_is_number(char *s)
 {
-    int i;
+	int i;
 
-    if (!s)
-        return (0);
-    i = 0;
-    while (s[i] && (s[i] == ' ' || s[i] == '+' || s[i] == '-'))
-    {
-        if (ft_isdigit(s[i]))
-            break ;
-        i++;
-        if (s[i] == '+' || s[i] == '-')
-            break ;
-    }
-    while (s[i])
-    {
-        if (!ft_isdigit(s[i]))
-            return (0);
-            i++;
-    }
-    return (1);
+	if (!s)
+		return (0);
+	i = 0;
+	while (s[i] && (s[i] == ' ' || s[i] == '+' || s[i] == '-'))
+	{
+		if (ft_isdigit(s[i]))
+			break ;
+		i++;
+		if (s[i] == '+' || s[i] == '-')
+			break ;
+	}
+	while (s[i])
+	{
+		if (!ft_isdigit(s[i]))
+			return (0);
+			i++;
+	}
+	return (1);
 }
 
 int test_duplicates(char **av)
 {
-    int i;
-    int j;
-    int len;
-    
-    i = 0;
-    while (av[i])
-    {
-        j = 0;
-        while (av[j])
-        {
-            len = ft_strlen(av[j]);
-            if (ft_strlen(av[i]) > len)
-                len = ft_strlen(av[i]);
-            if (!ft_strncmp(av[i], av[j], len) && i != j)
-                return 1;
-            j++;
-        }
-        i++;
-    }
-    return (0);
+	int i;
+	int j;
+	int len;
+	
+	i = 0;
+	while (av[i])
+	{
+		j = 0;
+		while (av[j])
+		{
+			len = ft_strlen(av[j]);
+			if (ft_strlen(av[i]) > (size_t)len)
+				len = ft_strlen(av[i]);
+			if (!ft_strncmp(av[i], av[j], len) && i != j)
+				return 1;
+			j++;
+		}
+		i++;
+	}
+	return (0);
 }
 
 t_stack load_stack(char *av[],int ac, int *flag)
 {
-    t_stack     a;
-    int         i;
-    long long   n;
+	t_stack     a;
+	int         i;
+	long long   n;
 
-    a.list = NULL;
-    *flag = 0;
-    i = ac;
-    *flag = test_duplicates(av);
-    while (--i > 0 && !*flag)
-    {
-        if (!ft_is_number(av[i]))
-        {
-            *flag = 1;
-            break ;
-        }
-        n = ft_atol(av[i]);
-        if (n > MAX_INT || n < MIN_INT || ft_strlen(av[i]) > 13)
-        {
-            *flag = 1;
-            break ;
-        }
-        a = push(a, n);
-    }
-    a.count = ac - 1;
-    return (a);
+	a.list = NULL;
+	*flag = 0;
+	i = ac;
+	*flag = test_duplicates(av);
+	while (--i > 0 && !*flag)
+	{
+		if (!ft_is_number(av[i]))
+		{
+			*flag = 1;
+			break ;
+		}
+		n = ft_atol(av[i]);
+		if (n > MAX_INT || n < MIN_INT || ft_strlen(av[i]) > 13)
+		{
+			*flag = 1;
+			break ;
+		}
+		a = push(a, n);
+	}
+	a.count = ac - 1;
+	return (a);
 }
